refactor(registrar): extract data-loading functions to useRegistrarData hook (Decomp 4)
Strategic Direction 1 — Decomposition (audit §8).

Extracted from RegistrarPanel.jsx (lines 540-766, ~226 lines) into a
dedicated custom hook. Three data-loading functions moved:

- loadIntegratedData: parallel fetch of doctors + services + departments
  from /registrar/* endpoints. Sets state via provided setters.
- fetchPatientData: fetches single patient by ID from /api/v1/patients/:id.
  Returns null for demo patients (ID >= 1000) or on network error.
- enrichAppointmentsWithPatientData: enriches appointment records with
  patient display fields (FIO, phone, birth year, gender, address) if
  missing from backend response. Also applies local overrides from
  appointmentOverridesRef.

NOT extracted (remain inline in RegistrarPanel):
- loadAppointments: depends on loadAppointmentsInFlightRef,
  autoRefreshCooldownUntilRef, autoRefreshCooldownLoggedRef,
  filteredAppointmentsRef — too tightly coupled to orchestrator state.
- loadMoreAppointments: depends on paginationInfo + loadAppointments.

Hook API:
  useRegistrarData({ setDoctors, setServices, setDynamicDepartments,
    appointmentOverridesRef })
  Returns: { loadIntegratedData, fetchPatientData,
    enrichAppointmentsWithPatientData }

Cleanup:
- Removed unused imports from RegistrarPanel.jsx:
  hasBackendPatientDisplayContract, hasBackendPatientGenderContract
  (now only used inside useRegistrarData.js)
- Removed formatNetworkErrorMessage, isNetworkFetchError imports
  (now only used inside useRegistrarData.js)
- Removed fetchPatientData from destructuring (not used in orchestrator)
- Updated contract test: added useRegistrarData.js to source tree,
  changed enrichAppointmentsWithPatientData block end marker from
  'const loadAppointments' (now in different file) to
  'return enrichedAppointments;' (last line of the function)

Stats:
- RegistrarPanel.jsx: 3521 → 3313 lines (−208 lines)
- useRegistrarData.js: 266 lines (new)
- Tests: 467/467 passing (no regression)
- ESLint: 0 errors (2 pre-existing warnings remain)
- Vite build: success
